pyproject.toml
[project]
name = "e2e9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.2",
    "pyyaml>=6.0.3",
    "ruff>=0.15.0",
    "soundfile>=0.13.1",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchcodec>=0.10.0",
]
